+++ src/Chronos.P2P.Client/Client/Model/ConnectHandshakeDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Chronos.P2P.Client
+{
+    internal struct ConnectHandshakeDto
+    {
+        public PeerEP Ep { get; set; }
+        public PeerInfo Info { get; set; }
+    }
+}
