+++ src/Chronos.P2P.Client/Client/Model/ConnectionReplyDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Chronos.P2P.Client
+{
+    internal struct ConnectionReplyDto
+    {
+        public bool Acc { get; set; }
+        public PeerEP Ep { get; set; }
+    }
+}
