+++ src/Chronos.P2P.Client/Client/Peer.cs
@@ -12,35 +12,12 @@ using System.Net.Sockets;
 using System.Text;
 using System.Text.Json;
 using System.Threading;
-using System.Threading.Channels;
 using System.Threading.Tasks;
 using static Chronos.P2P.Client.Utils;
 
 namespace Chronos.P2P.Client
 {
-    public class DatasliceSender
-    {
-        Guid sessionId;
-        Action<DataSlice>? sendAction;
-        internal void SetUp(Action<DataSlice> action, Guid guid)
-        {
-            sendAction = action;
-            sessionId = guid;
-        }
 
-        public void Send(byte[] data, int len)
-        {
-            var slice = new DataSlice
-            {
-                No = -1,
-                Slice = data,
-                Len = len,
-                Last = false,
-                SessionId = sessionId
-            };
-            sendAction!(slice);
-        }
-    }
     public class Peer : IRequestHandlerCollection, IDisposable
     {
         #region Fields
@@ -60,6 +37,8 @@ namespace Chronos.P2P.Client
         private CancellationTokenSource tokenSource = new();
         private long totalMsg = 0;
         private UdpClient udpClient;
+        private TaskCompletionSource<bool> connectionHandshakeTask 
+            = new(TaskCreationOptions.RunContinuationsAsynchronously);
         internal const int bufferLen = 40000;
         internal ConcurrentDictionary<Guid, FileRecvDicData> FileRecvDic = new();
         internal Stream? fs;
@@ -68,6 +47,8 @@ namespace Chronos.P2P.Client
         #endregion Fields
 
         #region Delegates & Events
+
+        public Func<PeerInfo, bool>? OnPeerInvited;
 
         public Func<BasicFileInfo, Task<(bool receive, string savePath)>>? OnInitFileTransfer;
 
@@ -132,73 +113,6 @@ namespace Chronos.P2P.Client
                         Ep = serverEP
                     });
                     await Task.Delay(1000, tokenSource.Token);
-                }
-            });
-
-        internal Task StartHolePunching()
-            => Task.Run(async () =>
-            {
-                if (peer!.OuterEP.IP == OuterEp!.IP)
-                {
-                    foreach (var item in LocalEP)
-                    {
-                        foreach (var item1 in peer.InnerEP)
-                        {
-                            if (item.IsInSameSubNet(item1))
-                            {
-                                peer.OuterEP = item1;
-                            }
-                        }
-                    }
-                }
-                int i = 0;
-                while (true)
-                {
-                    if (i>5)
-                    {
-                        i = 0;
-                        foreach (var item in LocalEP)
-                        {
-                            foreach (var item1 in peer!.InnerEP)
-                            {
-                                if (item.IsInSameSubNet(item1))
-                                {
-                                    peer.OuterEP = item1;
-                                    peer.InnerEP.Remove(item1);
-                                    Console.WriteLine($"trying new ep {item1}");
-                                    goto punch;
-                                }
-                            }
-                        }
-                    }
-                punch:
-                    if (peer is not null)
-                    {
-                        if (IsPeerConnected)
-                        {
-                            break;
-                        }
-                        if (tokenSource.IsCancellationRequested)
-                        {
-                            Console.WriteLine($"Connected data sent to peer {peer.OuterEP.ToIPEP()}");
-                            await SendDataToPeerAsync((int)CallMethods.Connected, "");
-                            if (IsPeerConnected)
-                            {
-                                break;
-                            }
-                        }
-                        else
-                        {
-                            Console.WriteLine($"Punching data sent to peer {peer.OuterEP.ToIPEP()}");
-                            await SendDataToPeerAsync((int)CallMethods.PunchHole, "");
-                            i++;
-                        }
-                        await Task.Delay(500);
-                    }
-                    else
-                    {
-                        await Task.Delay(1000);
-                    }
                 }
             });
 
@@ -276,6 +190,73 @@ namespace Chronos.P2P.Client
                     }
                 }
                 await Task.WhenAll(server.StartServerAsync(), StartPing(), StartPingWaiting());
+            });
+
+        internal Task StartHolePunching()
+                                    => Task.Run(async () =>
+            {
+                if (peer!.OuterEP.IP == OuterEp!.IP)
+                {
+                    foreach (var item in LocalEP)
+                    {
+                        foreach (var item1 in peer.InnerEP)
+                        {
+                            if (item.IsInSameSubNet(item1))
+                            {
+                                peer.OuterEP = item1;
+                            }
+                        }
+                    }
+                }
+                int i = 0;
+                while (true)
+                {
+                    if (i > 5)
+                    {
+                        i = 0;
+                        foreach (var item in LocalEP)
+                        {
+                            foreach (var item1 in peer!.InnerEP)
+                            {
+                                if (item.IsInSameSubNet(item1))
+                                {
+                                    peer.OuterEP = item1;
+                                    peer.InnerEP.Remove(item1);
+                                    Console.WriteLine($"trying new ep {item1}");
+                                    goto punch;
+                                }
+                            }
+                        }
+                    }
+                punch:
+                    if (peer is not null)
+                    {
+                        if (IsPeerConnected)
+                        {
+                            break;
+                        }
+                        if (tokenSource.IsCancellationRequested)
+                        {
+                            Console.WriteLine($"Connected data sent to peer {peer.OuterEP.ToIPEP()}");
+                            await SendDataToPeerAsync((int)CallMethods.Connected, "");
+                            if (IsPeerConnected)
+                            {
+                                break;
+                            }
+                        }
+                        else
+                        {
+                            Console.WriteLine($"Punching data sent to peer {peer.OuterEP.ToIPEP()}");
+                            await SendDataToPeerAsync((int)CallMethods.PunchHole, "");
+                            i++;
+                        }
+                        await Task.Delay(500);
+                    }
+                    else
+                    {
+                        await Task.Delay(1000);
+                    }
+                }
             });
 
         #endregion Workers
@@ -403,6 +384,23 @@ namespace Chronos.P2P.Client
             });
         }
 
+        public async Task<bool> RequestSendLiveStreamAsync(DatasliceSender sender, string name, int callMethod, CancellationToken token = default)
+        {
+            var sessionId = Guid.NewGuid();
+            sender.SetUp(slice =>
+            {
+                SendDataToPeerAsync(callMethod, slice);
+            }, sessionId);
+            FileAcceptTasks[sessionId] = new TaskCompletionSource<bool>(TaskCreationOptions.RunContinuationsAsynchronously);
+            await SendDataToPeerReliableAsync((int)CallMethods.StreamHandShake, new BasicFileInfo
+            {
+                Length = -1,
+                Name = name,
+                SessionId = sessionId
+            });
+            return await FileAcceptTasks[sessionId].Task;
+        }
+
         public async Task SendFileAsync(string location, int concurrentLevel = 3)
         {
             using SemaphoreSlim semaphore = new(concurrentLevel);
@@ -466,56 +464,10 @@ namespace Chronos.P2P.Client
             Console.WriteLine($"Auto adjusted timeout: {sendTimeOut}ms");
         }
 
-        public async Task<bool> RequestSendLiveStreamAsync(DatasliceSender sender, string name, int callMethod, CancellationToken token = default)
-        {
-            var sessionId = Guid.NewGuid();
-            sender.SetUp(slice =>
-            {
-                SendDataToPeerAsync(callMethod, slice);
-            }, sessionId);
-            FileAcceptTasks[sessionId] = new TaskCompletionSource<bool>(TaskCreationOptions.RunContinuationsAsynchronously);
-            await SendDataToPeerReliableAsync((int)CallMethods.StreamHandShake, new BasicFileInfo
-            {
-                Length = -1,
-                Name = name,
-                SessionId = sessionId
-            });
-            return await FileAcceptTasks[sessionId].Task;
-
-
-        }
-
         #endregion File Transfer
 
         #region Handlers
-        internal void OnConnectionRequested(PeerInfo requester)
-        {
-            var re = OnPeerInvited?.Invoke(requester);
-            if (!re.HasValue||re.Value)
-            {
-                Console.WriteLine("accept!");
-                peer = requester;
 
-                _ = SendDataReliableAsync((int)CallMethods.ConnectionHandShakeReply, new ConnectionReply
-                {
-                    Ep = requester.OuterEP,
-                    Acc = true
-                }, serverEP);
-            }
-            else
-            {
-                _ = SendDataReliableAsync((int)CallMethods.ConnectionHandShakeReply, new ConnectionReply
-                {
-                    Ep = requester.OuterEP,
-                    Acc = false
-                }, serverEP);
-            }
-        }
-        public Func<PeerInfo, bool>? OnPeerInvited;
-        internal void OnConnectionCallback(bool acc)
-        {
-            connectionHandshakeTask.TrySetResult(acc);
-        }
 
         private void Server_OnError(object? sender, byte[] e)
         {
@@ -535,6 +487,35 @@ namespace Chronos.P2P.Client
             if (AckTasks.TryGetValue(reqId, out var src))
             {
                 src.TrySetResult(true);
+            }
+        }
+
+        internal void OnConnectionCallback(bool acc)
+        {
+            connectionHandshakeTask.TrySetResult(acc);
+        }
+
+        internal void OnConnectionRequested(PeerInfo requester)
+        {
+            var re = OnPeerInvited?.Invoke(requester);
+            if (!re.HasValue || re.Value)
+            {
+                Console.WriteLine("accept!");
+                peer = requester;
+
+                _ = SendDataReliableAsync((int)CallMethods.ConnectionHandShakeReply, new ConnectionReplyDto
+                {
+                    Ep = requester.OuterEP,
+                    Acc = true
+                }, serverEP);
+            }
+            else
+            {
+                _ = SendDataReliableAsync((int)CallMethods.ConnectionHandShakeReply, new ConnectionReplyDto
+                {
+                    Ep = requester.OuterEP,
+                    Acc = false
+                }, serverEP);
             }
         }
 
@@ -627,15 +608,6 @@ namespace Chronos.P2P.Client
 
         #region Send Data
 
-        public Task SendDataToPeerAsync<T>(T data) where T : class
-        {
-            return SendDataToPeerAsync((int)CallMethods.P2PDataTransfer, data);
-        }
-
-        public Task SendDataToPeerAsync<T>(int method, T data)
-        {
-            return SendDataAsync(method, data, peer!.OuterEP.ToIPEP());
-        }
         public Task SendDataAsync<T>(int method, T data, IPEndPoint ep)
         {
             var t = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
@@ -653,15 +625,6 @@ namespace Chronos.P2P.Client
             return t.Task;
         }
 
-        public virtual ValueTask<bool> SendDataToPeerReliableAsync<T>(T data, CancellationToken? token = null)
-        {
-            return SendDataToPeerReliableAsync((int)CallMethods.P2PDataTransfer, data, 3, token);
-        }
-
-        public virtual ValueTask<bool> SendDataToPeerReliableAsync<T>(int method, T data, int retry = 10, CancellationToken? token = null)
-        {
-            return SendDataReliableAsync(method, data, peer!.OuterEP.ToIPEP(), retry, token);
-        }
         public virtual async ValueTask<bool> SendDataReliableAsync<T>(int method, T data, IPEndPoint ep, int retry = 10, CancellationToken? token = null)
         {
             Stopwatch? stopwatch = null;
@@ -726,12 +689,33 @@ namespace Chronos.P2P.Client
             return false;
         }
 
+        public Task SendDataToPeerAsync<T>(T data) where T : class
+        {
+            return SendDataToPeerAsync((int)CallMethods.P2PDataTransfer, data);
+        }
+
+        public Task SendDataToPeerAsync<T>(int method, T data)
+        {
+            return SendDataAsync(method, data, peer!.OuterEP.ToIPEP());
+        }
+
+        public virtual ValueTask<bool> SendDataToPeerReliableAsync<T>(T data, CancellationToken? token = null)
+        {
+            return SendDataToPeerReliableAsync((int)CallMethods.P2PDataTransfer, data, 3, token);
+        }
+
+        public virtual ValueTask<bool> SendDataToPeerReliableAsync<T>(int method, T data, int retry = 10, CancellationToken? token = null)
+        {
+            return SendDataReliableAsync(method, data, peer!.OuterEP.ToIPEP(), retry, token);
+        }
+
         #endregion Send Data
 
         #region User Interface
 
+
         public static Peer BuildWithStartUp<T>(int port, IPEndPoint serverEP, string? name = null)
-            where T : IStartUp, new()
+                    where T : IStartUp, new()
         {
             var peer = new Peer(port, serverEP, name);
             var startUp = new T();
@@ -739,8 +723,7 @@ namespace Chronos.P2P.Client
             peer.ConfigureServices(startUp.ConfigureServices);
             return peer;
         }
-        public void ConfigureServices(Action<ServiceCollection> configureAction)
-            => server.ConfigureServices(configureAction);
+
         public void AddHandler<T>() where T : class
             => server.AddHandler<T>();
 
@@ -748,6 +731,9 @@ namespace Chronos.P2P.Client
         {
             lifeTokenSource.Cancel();
         }
+
+        public void ConfigureServices(Action<ServiceCollection> configureAction)
+                            => server.ConfigureServices(configureAction);
 
         public void Dispose()
         {
@@ -769,7 +755,7 @@ namespace Chronos.P2P.Client
             await connectionHandshakeTask.Task;
             Console.WriteLine($"Trying remote ep: {peer.OuterEP}");
         }
-        TaskCompletionSource<bool> connectionHandshakeTask = new(TaskCreationOptions.RunContinuationsAsynchronously);
+
         public Task StartPeer()
         {
             var t = StartReceiveData();
@@ -779,16 +765,5 @@ namespace Chronos.P2P.Client
         }
 
         #endregion User Interface
-    }
-    internal struct ConnectionReply
-    {
-        public PeerEP Ep { get; set; }
-        public bool Acc { get; set; }
-
-    }
-    internal struct ConnectHandshakeDto
-    {
-        public PeerEP Ep { get; set; }
-        public PeerInfo Info { get; set; }
     }
 }
