+++ src/Chronos.P2P.Client/Server/Interface/IStartUp.cs
@@ -4,7 +4,8 @@ namespace Chronos.P2P.Server
 {
     public interface IStartUp
     {
-        public void ConfigureServices(IServiceCollection services);
         public void Configure(IRequestHandlerCollection handlers);
+
+        public void ConfigureServices(IServiceCollection services);
     }
 }
