+++ src/Chronos.P2P.Client/Server/P2PServer.cs
@@ -19,11 +19,11 @@ namespace Chronos.P2P.Server
         private UdpClient listener;
         private ConcurrentDictionary<Guid, PeerInfo> peers;
         private ServiceProvider? serviceProvider;
-        internal ServiceCollection services;
+        internal ConcurrentDictionary<PeerEP, (PeerEP, DateTime)> connectionDic = new();
         internal ConcurrentDictionary<Guid, DateTime> guidDic = new();
         internal MsgQueue<UdpMsg> msgs = new();
         internal Dictionary<int, TypeData> requestHandlers;
-        internal ConcurrentDictionary<PeerEP, (PeerEP, DateTime)> connectionDic = new();
+        internal ServiceCollection services;
 
         public event EventHandler? AfterDataHandled;
 
@@ -52,21 +52,6 @@ namespace Chronos.P2P.Server
                     msg.SendTask.SetResult();
                 }
             });
-        }
-
-        public static P2PServer BuildWithStartUp<T>(int port = 5000)
-            where T : IStartUp, new()
-        {
-            return BuildWithStartUp<T>(new UdpClient(new IPEndPoint(IPAddress.Any, port)));
-        }
-        public static P2PServer BuildWithStartUp<T>(UdpClient client)
-            where T : IStartUp, new()
-        {
-            var server = new P2PServer(client);
-            var startUp = new T();
-            startUp.Configure(server);
-            server.ConfigureServices(startUp.ConfigureServices);
-            return server;
         }
 
         /// <summary>
@@ -127,6 +112,22 @@ namespace Chronos.P2P.Server
             }
             CallHandler(td, new UdpContext(re.Buffer, peers, re.RemoteEndPoint, listener));
             AfterDataHandled?.Invoke(this, new());
+        }
+
+        public static P2PServer BuildWithStartUp<T>(int port = 5000)
+                                    where T : IStartUp, new()
+        {
+            return BuildWithStartUp<T>(new UdpClient(new IPEndPoint(IPAddress.Any, port)));
+        }
+
+        public static P2PServer BuildWithStartUp<T>(UdpClient client)
+            where T : IStartUp, new()
+        {
+            var server = new P2PServer(client);
+            var startUp = new T();
+            startUp.Configure(server);
+            server.ConfigureServices(startUp.ConfigureServices);
+            return server;
         }
 
         /// <summary>
